test: replace azurite and fake gcs helpers with floci
Rename the Azure and Google storage testcontainer modules to Floci-specific helpers and switch their runtime images to floci-az and floci-gcp.

This updates Quarkus catalog test resources, GC object-storage integration tests, BOM constraints, and Gradle project wiring so no test helper module remains named after Azurite or fake-gcs.
