+++ bom/build.gradle.kts
@@ -26,7 +26,6 @@ publishingHelper { mavenName = "Nessie - Bill of Materials (BOM)" }
 dependencies {
   constraints {
     api(project(":"))
-    api(project(":nessie-azurite-testcontainer"))
     api(project(":nessie-cli"))
     api(project(":nessie-cli-grammar"))
     api(project(":nessie-client"))
@@ -43,15 +42,16 @@ dependencies {
     api(project(":nessie-events-service"))
     api(project(":nessie-events-quarkus"))
     api(project(":nessie-events-ri"))
+    api(project(":nessie-floci-az-testcontainer"))
+    api(project(":nessie-floci-gcp-testcontainer"))
+    api(project(":nessie-floci-s3-testcontainer"))
     api(project(":nessie-gc-base"))
     api(project(":nessie-gc-repository-jdbc"))
-    api(project(":nessie-gcs-testcontainer"))
     api(project(":nessie-model"))
     api(project(":nessie-notice"))
     api(project(":nessie-jaxrs-testextension"))
     api(project(":nessie-jaxrs-tests"))
     api(project(":nessie-keycloak-testcontainer"))
-    api(project(":nessie-floci-s3-testcontainer"))
     api(project(":nessie-nessie-testcontainer"))
     api(project(":nessie-network-tools"))
     api(project(":nessie-object-storage-mock"))
