+++ gradle/projects.main.properties
@@ -1,6 +1,5 @@
 nessie-aggregated-license-report=tools/aggregated-license-report
 nessie-notice=tools/notice
-nessie-azurite-testcontainer=testing/azurite-container
 nessie-bom=bom
 nessie-cli=cli/cli
 nessie-cli-grammar=cli/grammar
@@ -23,7 +22,6 @@ nessie-model=api/model
 nessie-gc-base=gc/gc-base
 nessie-gc-base-tests=gc/gc-base-tests
 nessie-gc-repository-jdbc=gc/gc-repository-jdbc
-nessie-gcs-testcontainer=testing/gcs-container
 nessie-immutables=tools/immutables
 nessie-immutables-std=tools/immutables-std
 nessie-perftest-gatling=perftest/gatling
@@ -34,6 +32,8 @@ nessie-keycloak-testcontainer=testing/keycloak-container
 nessie-nessie-testcontainer=testing/nessie-container
 nessie-network-tools=tools/network
 nessie-object-storage-mock=testing/object-storage-mock
+nessie-floci-az-testcontainer=testing/floci-az-container
+nessie-floci-gcp-testcontainer=testing/floci-gcp-container
 nessie-floci-s3-testcontainer=testing/floci-s3-container
 nessie-quarkus=servers/quarkus-server
 nessie-quarkus-authn=servers/quarkus-authn
