+++ servers/quarkus-server/build.gradle.kts
@@ -150,8 +150,8 @@ dependencies {
   testFixturesApi("org.testcontainers:testcontainers")
   testFixturesApi(project(":nessie-keycloak-testcontainer"))
   testFixturesApi("org.testcontainers:testcontainers-vault")
-  testFixturesApi(project(":nessie-azurite-testcontainer"))
-  testFixturesApi(project(":nessie-gcs-testcontainer"))
+  testFixturesApi(project(":nessie-floci-az-testcontainer"))
+  testFixturesApi(project(":nessie-floci-gcp-testcontainer"))
   testFixturesApi(project(":nessie-floci-s3-testcontainer"))
   testFixturesApi(project(":nessie-trino-testcontainer"))
   testFixturesApi(project(":nessie-object-storage-mock"))
