--- testing/azurite-container/src/main/resources/org/projectnessie/testing/azurite/Dockerfile-azurite-version
@@ -1,3 +1,0 @@
-# Dockerfile to provide the image name and tag to a test.
-# Version is managed by Renovate - do not edit.
-FROM mcr.microsoft.com/azure-storage/azurite:3.36.0
