+++ testing/floci-az-container/src/main/java/org/projectnessie/testing/floci/az/FlociAzExtension.java
@@ -1,0 +1,123 @@
+/*
+ * Copyright (C) 2024 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.testing.floci.az;
+
+import static java.lang.String.format;
+import static org.junit.jupiter.api.extension.ConditionEvaluationResult.disabled;
+import static org.junit.jupiter.api.extension.ConditionEvaluationResult.enabled;
+import static org.junit.platform.commons.util.AnnotationUtils.findAnnotatedFields;
+import static org.junit.platform.commons.util.ReflectionUtils.makeAccessible;
+
+import java.lang.reflect.Field;
+import org.junit.jupiter.api.condition.OS;
+import org.junit.jupiter.api.extension.BeforeAllCallback;
+import org.junit.jupiter.api.extension.BeforeEachCallback;
+import org.junit.jupiter.api.extension.ConditionEvaluationResult;
+import org.junit.jupiter.api.extension.ExecutionCondition;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.api.extension.ParameterResolutionException;
+import org.junit.jupiter.api.extension.ParameterResolver;
+import org.junit.platform.commons.util.AnnotationUtils;
+import org.junit.platform.commons.util.ExceptionUtils;
+import org.junit.platform.commons.util.ReflectionUtils;
+
+public class FlociAzExtension
+    implements BeforeAllCallback, BeforeEachCallback, ParameterResolver, ExecutionCondition {
+  private static final ExtensionContext.Namespace NAMESPACE =
+      ExtensionContext.Namespace.create(FlociAzExtension.class);
+
+  @Override
+  public ConditionEvaluationResult evaluateExecutionCondition(ExtensionContext context) {
+    if (OS.current() == OS.LINUX) {
+      return enabled("Running on Linux");
+    }
+    if (OS.current() == OS.MAC) {
+      return enabled("Running on macOS");
+    }
+    return disabled(format("Disabled on %s", OS.current().name()));
+  }
+
+  @Override
+  public void beforeAll(ExtensionContext context) {
+    Class<?> testClass = context.getRequiredTestClass();
+
+    findAnnotatedFields(testClass, FlociAz.class, ReflectionUtils::isStatic)
+        .forEach(field -> injectField(context, field));
+  }
+
+  @Override
+  public void beforeEach(ExtensionContext context) {
+    Class<?> testClass = context.getRequiredTestClass();
+
+    findAnnotatedFields(testClass, FlociAz.class, ReflectionUtils::isNotStatic)
+        .forEach(field -> injectField(context, field));
+  }
+
+  private void injectField(ExtensionContext context, Field field) {
+    try {
+      FlociAz flociAz =
+          AnnotationUtils.findAnnotation(field, FlociAz.class)
+              .orElseThrow(IllegalStateException::new);
+
+      FlociAzAccess container =
+          context
+              .getStore(NAMESPACE)
+              .computeIfAbsent(
+                  field.toString(), x -> createContainer(flociAz), FlociAzAccess.class);
+
+      makeAccessible(field).set(context.getTestInstance().orElse(null), container);
+    } catch (Throwable t) {
+      ExceptionUtils.throwAsUncheckedException(t);
+    }
+  }
+
+  @Override
+  public boolean supportsParameter(
+      ParameterContext parameterContext, ExtensionContext extensionContext)
+      throws ParameterResolutionException {
+    if (parameterContext.findAnnotation(FlociAz.class).isEmpty()) {
+      return false;
+    }
+    return parameterContext.getParameter().getType().isAssignableFrom(FlociAzAccess.class);
+  }
+
+  @Override
+  public Object resolveParameter(
+      ParameterContext parameterContext, ExtensionContext extensionContext)
+      throws ParameterResolutionException {
+    return extensionContext
+        .getStore(NAMESPACE)
+        .computeIfAbsent(
+            FlociAzExtension.class.getName() + '#' + parameterContext.getParameter().getName(),
+            k -> {
+              FlociAz flociAz = parameterContext.findAnnotation(FlociAz.class).get();
+              return createContainer(flociAz);
+            },
+            FlociAzAccess.class);
+  }
+
+  private FlociAzAccess createContainer(FlociAz flociAz) {
+    String bucket = nonDefault(flociAz.storageContainer());
+    FlociAzContainer container = new FlociAzContainer(null, bucket).withStartupAttempts(5);
+    container.start();
+    return container;
+  }
+
+  private static String nonDefault(String s) {
+    return s.equals(FlociAz.DEFAULT) ? null : s;
+  }
+}
