+++ testing/floci-az-container/src/main/resources/org/projectnessie/testing/floci/az/Dockerfile-floci-az-version
@@ -1,0 +1,3 @@
+# Dockerfile to provide the image name and tag to a test.
+# Version is managed by Renovate - do not edit.
+FROM docker.io/floci/floci-az:0.9.0
