+++ testing/floci-gcp-container/src/main/java/org/projectnessie/testing/floci/gcp/FlociGcp.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2022 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.testing.floci.gcp;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.FIELD, ElementType.PARAMETER})
+@Retention(RetentionPolicy.RUNTIME)
+@Inherited
+public @interface FlociGcp {
+  /** Optional, use this project ID instead of a random one. */
+  String projectId() default DEFAULT;
+
+  /** Optional, use this OAuth 2 token instead of a random one. */
+  String oauth2token() default DEFAULT;
+
+  /** Optional, use this bucket instead of a random one. */
+  String bucket() default DEFAULT;
+
+  String DEFAULT = "floci_gcp_default_value__";
+}
