+++ testing/floci-gcp-container/src/main/resources/org/projectnessie/testing/floci/gcp/Dockerfile-floci-gcp-version
@@ -1,0 +1,3 @@
+# Dockerfile to provide the image name and tag to a test.
+# Version is managed by Renovate - do not edit.
+FROM docker.io/floci/floci-gcp:0.5.0
