--- testing/gcs-container/src/main/resources/org/projectnessie/testing/gcs/Dockerfile-fake-gcs-server-version
@@ -1,3 +1,0 @@
-# Dockerfile to provide the image name and tag to a test.
-# Version is managed by Renovate - do not edit.
-FROM docker.io/fsouza/fake-gcs-server:1.55.0
